[Diamond V] Title: 가장 큰 원, Time: 636 ms, Memory: 2172 KB -BaekjoonHub

+++ 백준/Diamond/3726. 가장 큰 원/가장 큰 원.cc
@@ -1,6 +1,6 @@
 #include <bits/stdc++.h>
 using namespace std;
-using rea=long double;
+using rea=double;
 constexpr rea eps=1e-12L;
 
 struct Seg {
